refactor(mocha): split scanner.rs into statements, expressions, and checks
